+++ packages/example-antd-management-fast-framework/src/businessComponents/Drawers/Single/SelectDrawer/FrontendPagination/SimpleSinglePageFrontendPaginationMultiSelectDrawer/codeSource.js
@@ -26,6 +26,7 @@ class SimpleSinglePageFrontendPaginationMultiSelectDrawer extends BaseSimpleSing
 
     this.state = {
       ...this.state,
+      showSelect: true,
       sourceCode: code,
     };
   }