+++ packages/example-antd-management-fast-framework/src/businessComponents/Modals/Multi/SelectModal/SimpleMultiPageMultiSelectModal/index.jsx
@@ -1,6 +1,5 @@
 import { connect } from 'easy-soft-dva';
 
-import { logTemplate } from 'antd-management-fast-common';
 import { switchControlAssist } from 'antd-management-fast-framework';
 
 import BaseSimpleMultiPageSelectModal from '../BaseSimpleMultiPageSelectModal';
@@ -28,12 +27,6 @@ class SimpleMultiPageMultiSelectModal extends BaseSimpleMultiPageSelectModal {
       sourceCode: code,
     };
   }
-
-  onBatchActionClick = (list) => {
-    logTemplate({
-      list,
-    });
-  };
 }
 
 export { SimpleMultiPageMultiSelectModal };