+++ packages/antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 2.11.112
+
+### Patch Changes
+
+- update
+
+## 2.11.111
+
+### Patch Changes
+
+- update
+
 ## 2.11.110
 
 ### Patch Changes